Reworked imports and moved Enums
