+++ environment.yml
@@ -5,4 +5,5 @@ dependencies:
   - python=3.7
   - flake8
   - pytest
+  - numpy
